# Improve code generation for homogenous tuples of unknown length