E.g.:
```
mutable struct foo
    x::Int
end
mutable struct bar
    tup::NTuple{N, foo} where N
end
function f(b::bar)
    sum = 0
    for i = 1:2
        @inbounds sum += b.tup[i].x
    end
    sum
end
global x = bar((foo(1),foo(2)))
@benchmark f(x::bar)
```

Before:
```
julia> @benchmark f(x::bar)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     27.124 ns (0.00% GC)
  median time:      28.771 ns (0.00% GC)
  mean time:        29.864 ns (0.00% GC)
  maximum time:     116.762 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     995
```

After:
```
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     10.531 ns (0.00% GC)
  median time:      10.559 ns (0.00% GC)
  mean time:        10.610 ns (0.00% GC)
  maximum time:     122.751 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999
  time tolerance:   5.00%
  memory tolerance: 1.00%
```