@nanosoldier 2017-04-04
[Your benchmark job](https://github.com/JuliaLang/julia/pull/21263#issuecomment-291293596) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/017dc389a5b6f754cf93e1948d2d10918dd80ea8/40c4860_vs_daefda4/report.md). cc @jrevels