@Keno 2017-04-06
A bunch of these results were also flakey in other logs. Let's try again to see if we have something real:

@nanosoldier `runbenchmarks(ALL, vs=":master")`