@nanosoldier 2017-04-06
[Your benchmark job](https://github.com/JuliaLang/julia/pull/21263#issuecomment-292301329) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/29618a797cd24e069f0509c7e080af18151f155a/165ec20_vs_d20c66d/report.md). cc @jrevels